% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretModels.R
\name{caretModels}
\alias{caretModels}
\title{Build a caret model}
\usage{
caretModels(TrainSet, resp.var, control, preProc.opt, tuneL, metric,
  methods)
}
\description{
A function to build prediction models for the training set.
kernlab requires formula instead of y vector and X matrix
}
