% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circbar.R
\name{Circbar}
\alias{Circbar}
\title{Create a circular bar plot}
\usage{
Circbar(mydata, graphtitle)
}
\arguments{
\item{mydata}{The dataset that you want to plot.}

\item{graphtitle}{The title of the plot.}
}
\description{
Was created to display var importance plots
}
