% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_rep.R
\name{comb_rep}
\alias{comb_rep}
\title{Combine results from parallelized replication (outer) loop}
\usage{
comb_rep(LL1, LL2)
}
\description{
A function to combine results from parallelized replication (outer) loop.
}
