% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.code.data.R
\name{dummy.code.data}
\alias{dummy.code.data}
\title{Convert categorical predictors to dummy coding}
\usage{
dummy.code.data(data, var.list, ref.list)
}
\description{
A function to convert categorical predictors to dummy coding.
}
