% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdNCV.R
\name{pdNCV}
\alias{pdNCV}
\title{Partial-depedence plots using repeated nested CV objects
Current PDP only allows stacking by weighted average.
In classification, stacking by other learner uses different "predictor names".}
\usage{
pdNCV(sel.var, resp.var, nRep, nFolds, dir.path, file.root, rncv.obj,
  stack.wt)
}
\description{
Partial-depedence plots using repeated nested CV objects
Current PDP only allows stacking by weighted average.
In classification, stacking by other learner uses different "predictor names".
}
