% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdNCV.to.be.fixed.R
\name{pdNCV.to.be.fixed}
\alias{pdNCV.to.be.fixed}
\title{No Description.}
\usage{
pdNCV.to.be.fixed(sel.var, resp.var, nRep, nFolds, dir.path, file.root,
  rncv.obj, stack.method, stack.wt)
}
\description{
No Description.
}
