% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest.rNCV.R
\name{PermTest.rNCV}
\alias{PermTest.rNCV}
\title{No Description.}
\usage{
PermTest.rNCV(rncv.obj, perf.perm, figure)
}
\description{
Uses results from permTest() for finding significance of model performance. This function may be time-consuming.
See lines 589 - 605 for an example.
}
