% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermuPerf.rNCV.R
\name{PermuPerf.rNCV}
\alias{PermuPerf.rNCV}
\title{No Description.}
\usage{
PermuPerf.rNCV(data, resp.var, ref.lv = NULL, nRep, nFolds.outer,
  ML.methods, control, tuneL, preProc.opt, metric, dir.path, file.root,
  stack.method = "wt.avg", weighted.by = NULL, stack.wt = NULL,
  control.stack = NULL, nPerm)
}
\description{
No Description.
}
