% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuPred.R
\name{permuPred}
\alias{permuPred}
\title{No Description.}
\usage{
permuPred(
  TrainSet,
  TestSet,
  resp.var,
  ref.lv = NULL,
  control,
  preProc.opt,
  tuneL,
  metric,
  methods,
  stack.method = "none",
  weighted.by = NULL,
  stack.wt = NULL,
  control.stack = NULL,
  select = "none",
  nPerm
)
}
\description{
No Description.
}
