% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredVal.R
\name{PredVal}
\alias{PredVal}
\title{No Description.}
\usage{
PredVal(models, TestSet, resp.var, ref.lv = NULL,
  stack.method = "none", weighted.by = NULL, stack.wt = NULL,
  control.stack = NULL, tuneL = NULL)
}
\description{
No Description.
}
