% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_cormat.R
\name{reorder_cormat}
\alias{reorder_cormat}
\title{A function to print similarity across ML predictions.}
\usage{
reorder_cormat(cormat)
}
\description{
A function to print similarity across ML predictions.
}
