% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNCV.perf.summ.R
\name{rNCV.perf.summ}
\alias{rNCV.perf.summ}
\title{Summarize model performance in rNCV}
\usage{
rNCV.perf.summ(rNCV.obj)
}
\arguments{
\item{rNCV.obj}{The rNCV object returned by the function \code{rNCV()}.}
}
\description{
Gives a summary of model performance. Returns average of each metric specified in rNCV (whether MAE, RMSE, R^2, or something else) for both the training and testing set for each method.
}
