% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNCV.R
\name{rNCV}
\alias{rNCV}
\title{No Description.}
\usage{
rNCV(data, resp.var, ref.lv = NULL, nRep, nFolds.outer, ML.methods,
  control, tuneL, preProc.opt, metric, dir.path, file.root,
  stack.method = "wt.avg", weighted.by = NULL, stack.wt = NULL,
  control.stack = NULL)
}
\description{
No Description.
}
