% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPred.R
\name{stackPred}
\alias{stackPred}
\title{Predict a new set of data using a caretStack model.}
\usage{
stackPred(
  data,
  resp.var,
  nRep,
  nFolds.outer,
  dir.path,
  file.root,
  stack.method = "wt.avg"
)
}
\arguments{
\item{data}{The data frame containing the training set.}

\item{resp.var}{Variable to predict.}

\item{nRep}{Number of times nCV is repeated.}

\item{nFolds.outer}{Number of outer folds}

\item{dir.path}{Directory where the CV data is stored. Example: "Output/rNCV/"}

\item{file.root}{Prefix for the CV filenames.}
}
\description{
Supports classification and regression.
Note: only continuous variables are expected to be used as predictors. It is assumed that there are a sufficient number of subjects in each category.
}
