% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Create a plot describing model performance}
\usage{
summary_plot(summ, figure_title, metrics = c("MAE", "RMSE", "Rsquared"))
}
\arguments{
\item{figure_title}{The title you would like for the figure}

\item{metrics}{Which performance measures you would like to use.}
}
\description{
Create a plot describing model performance
}
