% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainStack.R
\name{trainStack}
\alias{trainStack}
\title{Train a stacked model using caret}
\usage{
trainStack(x, y, layers, folds, verbose = F)
}
\arguments{
\item{x}{Predictors}

\item{y}{Response}

\item{layers}{Stack layers. A list of lists of models.}

\item{folds}{CV folds, as created by caret::createFolds}

\item{verbose}{Output progress messages}
}
\description{
Train a stacked model using caret
}
\examples{
stackLayers <- list(
 list(
   gbm2 = list(
     parallel = 4,
     params = list(
       method = "gbm",
       tuneGrid = expand.grid(
         n.trees = 300,
         interaction.depth = 2,
         shrinkage = 0.1,
         n.minobsinnode = 10
       ),
       trControl = trainControl(method = "none")
     )
   ),
   gbm10 = list(
     parallel = 4,
     params = list(
       method = "gbm",
       tuneGrid = expand.grid(
         n.trees = 300,
         interaction.depth = 10,
         shrinkage = 0.1,
         n.minobsinnode = 10
       ),
       trControl = trainControl(method = "none")
     )
   )
  ),
 list(
    gbm2 = list(
      parallel = 4,
      params = list(
        method = "gbm",
        tuneGrid = expand.grid(
          n.trees = 300,
          interaction.depth = 2,
          shrinkage = 0.1,
          n.minobsinnode = 10
        ),
        trControl = trainControl(method = "none")
      )
    )
)
)

folds <- caret::createFolds(x, 10)
stackModel <- trainStack(x, y, stackLayers, folds)
}
