% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainStack.R
\name{trainStackNoRestack}
\alias{trainStackNoRestack}
\title{Train a stacked model using caret, without restacking}
\usage{
trainStackNoRestack(x, y, layers, folds, verbose = F)
}
\arguments{
\item{x}{Predictors}

\item{y}{Response}

\item{layers}{Stack layers. A list of lists of models.}

\item{folds}{CV folds, as created by caret::createFolds}

\item{verbose}{Output progress messages}
}
\description{
Train a stacked model using caret, without restacking
}
