% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp_plot.R
\name{varimp_plot}
\alias{varimp_plot}
\title{Plots variable importance from ML model}
\usage{
varimp_plot(this_VarImp)
}
\arguments{
\item{this_VarImp}{Use the file exported from the rNCV function by the name '[your_prefix]_VarImp.csv'}
}
\description{
This function plots variable importance.
}
