% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp_rNCV.R
\name{varImp_rNCV}
\alias{varImp_rNCV}
\title{Variable importance for repeated nested CV.}
\usage{
varImp_rNCV(rNCV.obj)
}
\arguments{
\item{rNCV.obj}{The rNCV object returned by the function \code{rNCV()}.}
}
\description{
Variable importance for repeated nested CV.
}
