% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretModels.R
\name{caretModels}
\alias{caretModels}
\title{Build a caret model}
\usage{
caretModels(TrainSet, resp.var, trControl, preProcess, tuneLength, metric,
  methods)
}
\arguments{
\item{TrainSet}{The training set.}

\item{resp.var}{Indicate the name of the column in the training set that contains the response variable.}

\item{trControl}{A list of values that define how this function
acts. See \code{\link{trainControl}} and
\url{http://topepo.github.io/caret/using-your-own-model-in-train.html}.
(NOTE: If given, this argument must be named.)}

\item{preProcess}{A string vector that defines a pre-processing
of the predictor data. Current possibilities are "BoxCox",
"YeoJohnson", "expoTrans", "center", "scale", "range",
"knnImpute", "bagImpute", "medianImpute", "pca", "ica" and
"spatialSign". The default is no pre-processing. See
\code{\link{preProcess}} and \code{\link{trainControl}} on the
procedures and how to adjust them. Pre-processing code is only
designed to work when \code{x} is a simple matrix or data frame.}

\item{tuneLength}{An integer denoting the amount of granularity
in the tuning parameter grid. By default, this argument is the
number of levels for each tuning parameters that should be
generated by \code{\link{train}}. If \code{\link{trainControl}}
has the option \code{search = "random"}, this is the maximum
number of tuning parameter combinations that will be generated
by the random search. (NOTE: If given, this argument must be
named.)}

\item{metric}{A string that specifies what summary metric will
be used to select the optimal model. By default, possible values
are "RMSE" and "Rsquared" for regression and "Accuracy" and
"Kappa" for classification. If custom performance metrics are
used (via the \code{summaryFunction} argument in
\code{\link{trainControl}}, the value of \code{metric} should
match one of the arguments. If it does not, a warning is issued
and the first metric given by the \code{summaryFunction} is
used. (NOTE: If given, this argument must be named.)}

\item{methods}{Similarly to the \code{method} argument in caret's \code{train} function, this argument is a list of strings specifying which classification or regression models to use. Possible values are found using \code{names(getModelInfo())}. See \url{http://topepo.github.io/caret/train-models-by-tag.html}. A list of functions can also be passed for a custom model function. See \url{http://topepo.github.io/caret/using-your-own-model-in-train.html} for details.}
}
\description{
A function to build prediction models for the training set.
}
\examples{
ctrl.reg <- trainControl(method = 'cv',               # k-fold cross-validation
number = 5,                  # k = 5
search = 'grid',             # use grid search over paramter space
summaryFunction = defaultSummary,
selectionFunction = 'oneSE', # select optimal tuning parameters by "one standard error" rule
savePredictions = 'final')   # save predicted values of the final model

boston.models <- caretModels(boston.training,    # training set
                             resp.var='cmedv',   # response variable
                             trControl = ctrl.reg,
                             preProcess = c('center', 'scale'),
                             tuneLength = 7,
                             metric = 'RMSE',
                             methods = 'svmRadial')
}
