% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictStack.R
\name{predictStack}
\alias{predictStack}
\title{Predict using a stacked model}
\usage{
predictStack(model, x, verbose = F)
}
\arguments{
\item{model}{Result of running trainStack}

\item{x}{Predictors}
}
\description{
Predict using a stacked model
}
