% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNCV.R
\name{rNCV}
\alias{rNCV}
\title{Repeated, Nested Cross-Validation}
\usage{
rNCV(data, resp.var, ref.lv = NULL, nRep, nFolds.outer, methods,
  trControl, tuneLength, preProcess, metric, dir.path, file.root,
  stack.method = "wt.avg", weighted.by = NULL, stack.wt = NULL,
  control.stack = NULL)
}
\arguments{
\item{data}{The data frame containing the training set.}

\item{resp.var}{Indicate the name of the column in the training set that contains the response variable.}

\item{ref.lv}{???}

\item{nRep}{Number of times nCV is repeated.}

\item{nFolds.outer}{Number of outer folds}

\item{methods}{Similarly to the \code{method} argument in caret's \code{train} function, this argument is a list of strings specifying which classification or regression models to use. Possible values are found using \code{names(getModelInfo())}. See \url{http://topepo.github.io/caret/train-models-by-tag.html}. A list of functions can also be passed for a custom model function. See \url{http://topepo.github.io/caret/using-your-own-model-in-train.html} for details.}

\item{trControl}{A list of values that define how this function
acts. See \code{\link{trainControl}} and
\url{http://topepo.github.io/caret/using-your-own-model-in-train.html}.
(NOTE: If given, this argument must be named.)}

\item{tuneLength}{An integer denoting the amount of granularity
in the tuning parameter grid. By default, this argument is the
number of levels for each tuning parameters that should be
generated by \code{\link{train}}. If \code{\link{trainControl}}
has the option \code{search = "random"}, this is the maximum
number of tuning parameter combinations that will be generated
by the random search. (NOTE: If given, this argument must be
named.)}

\item{preProcess}{A string vector that defines a pre-processing
of the predictor data. Current possibilities are "BoxCox",
"YeoJohnson", "expoTrans", "center", "scale", "range",
"knnImpute", "bagImpute", "medianImpute", "pca", "ica" and
"spatialSign". The default is no pre-processing. See
\code{\link{preProcess}} and \code{\link{trainControl}} on the
procedures and how to adjust them. Pre-processing code is only
designed to work when \code{x} is a simple matrix or data frame.}

\item{metric}{A string that specifies what summary metric will
be used to select the optimal model. By default, possible values
are "RMSE" and "Rsquared" for regression and "Accuracy" and
"Kappa" for classification. If custom performance metrics are
used (via the \code{summaryFunction} argument in
\code{\link{trainControl}}, the value of \code{metric} should
match one of the arguments. If it does not, a warning is issued
and the first metric given by the \code{summaryFunction} is
used. (NOTE: If given, this argument must be named.)}

\item{dir.path}{Directory where the CV data is stored.}

\item{file.root}{Prefix for the CV filenames.}

\item{stack.method}{???}

\item{weighted.by}{???}

\item{stack.wt}{???}

\item{control.stack}{???}
}
\description{
Repeated, Nested Cross-Validation
}
