% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_one.R
\name{summarize_one}
\alias{summarize_one}
\title{Plot ML Results and Summary}
\usage{
summarize_one(file_name, figure_title, rdata_prefix, outDir = "",
  include_pdp = FALSE)
}
\arguments{
\item{figure_title}{The name you would like for the generated figures.}

\item{rdata_prefix}{Label of output file name. Make sure this is the same label as specified in \code{predict_one()}.}

\item{include_pdp}{Set to TRUE if you would like to generate partial dependence plots. \strong{WARNING:} Setting this option to TRUE will significantly increase run time.}

\item{data_folder}{The folder where the results from the rNCV are contained.}

\item{ourDir}{If you would like to save the output files into somewhere other than the working directory, specify that here. Make sure the folder name ends with '/'.}
}
\description{
Plot ML Results and Summary
}
